requirements-dev.txt
# Development dependencies for qbit-torrent-extract
# Install with: pip install -r requirements-dev.txt

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
psutil>=5.9.0  # For performance tests

# Code quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0
types-tqdm>=4.65.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.3.0
sphinx-click>=4.4.0

# Development tools
pre-commit>=3.0.0
ipython>=8.10.0  # For interactive debugging
ipdb>=0.13.13    # For debugging

# Build tools
build>=0.10.0
twine>=4.0.2     # For PyPI uploads
wheel>=0.40.0

# Security scanning
bandit>=1.7.5
safety>=2.3.5